dt_fill {tidytable}	R Documentation

Fill in missing values with previous or next value

Description

Fills missing values in the selected columns using the next or previous entry. Can be done by group.

Usage

dt_fill(
  .data,
  ...,
  .direction = c("down", "up", "downup", "updown"),
  by = NULL
)

Arguments

.data
A data.frame or data.table

...
A selection of bare columns

.direction
Direction in which to fill missing values. Currently "down" (the default), "up", "downup" (first down then up), or "updown" (first up and then down)

by
Whether the filling should be done by group. Passed in a list()

Examples

test_df <- data.table::data.table(
  x = c(NA, NA, NA, 4:10),
  y = c(1:6, NA, 8, NA, 10),
  z = c(rep("a", 8), rep("b", 2)))

test_df %>%
  dt_fill(x, y, by = z)

test_df %>%
  dt_fill(x, y, by = z, .direction = "downup")

[Package tidytable version 0.3.1 Index]