dt_left_join {tidytable}	R Documentation

Join two data.tables together

Description

Join two data.tables together

Usage

dt_left_join(x, y, by = NULL)

dt_inner_join(x, y, by = NULL)

dt_right_join(x, y, by = NULL)

dt_full_join(x, y, by = NULL, suffix = c(".x", ".y"))

dt_anti_join(x, y, by = NULL)

Arguments

x
A data.frame or data.table

y
A data.frame or data.table

by
A character vector of variables to join by. If NULL, the default, the join will do a natural join, using all variables with common names across the two tables.

suffix
Append created for duplicated column names when using 'dt_full_join()'

Value

A data.table

Examples

df1 <- data.table::data.table(x = c("a","a","a","b","b"), y = 1:5)
df2 <- data.table::data.table(x = c("a","b"), z = 1:2)

df1 %>% dt_left_join(df2)
df1 %>% dt_inner_join(df2)
df1 %>% dt_right_join(df2)
df1 %>% dt_full_join(df2)
df1 %>% dt_anti_join(df2)

[Package tidytable version 0.3.1 Index]