dt_map {tidytable}	R Documentation

Apply a function to each element of a vector

Description

The dt_map functions transform their input by applying a function to each element and returning a vector the same length as the input.

dt_map() returns a list

_lgl(), _int(), _dbl() and _chr() variants return their specified type

_dfr() & _dfc() Return data frame results binded together

Usage

dt_map(.x, .f, ...)

dt_map_lgl(.x, .f, ...)

dt_map_int(.x, .f, ...)

dt_map_dbl(.x, .f, ...)

dt_map_chr(.x, .f, ...)

dt_map_dfc(.x, .f, ...)

dt_map_dfr(.x, .f, ..., .id = NULL)

dt_walk(.x, .f, ...)

dt_map2(.x, .y, .f, ...)

dt_map2_lgl(.x, .y, .f, ...)

dt_map2_int(.x, .y, .f, ...)

dt_map2_dbl(.x, .y, .f, ...)

dt_map2_chr(.x, .y, .f, ...)

dt_map2_dfc(.x, .y, .f, ...)

dt_map2_dfr(.x, .y, .f, ..., .id = NULL)

Arguments

.x
A list or vector

.f
A function

...
Other arguments to pass to a function

.id
Whether dt_map_dfr() should add an id column to the finished dataset

.y
A list or vector

Examples

dt_map(c(1,2,3), ~.x + 1)

dt_map_dbl(c(1,2,3), ~.x + 1)

dt_map_chr(c(1,2,3), as.character)

[Package tidytable version 0.3.1 Index]