Fill in missing values with previous or next value

Fills missing values in the selected columns using the next or previous entry. Can be done by group.

Supports tidyselect

fill.(.df, ..., .direction = c("down", "up", "downup", "updown"), .by = NULL)

Arguments

.df
A data.frame or data.table

...
A selection of columns. tidyselect compatible.

.direction
Direction in which to fill missing values. Currently "down" (the default), "up", "downup" (first down then up), or "updown" (first up and then down)

.by
Columns to group by when filling should be done by group

Examples

test_df <- tidytable(
  x = c(NA, NA, NA, 4:10),
  y = c(1:6, NA, 8, NA, 10),
  z = c(rep("a", 8), rep("b", 2)))

test_df %>%
  fill.(x, y, .by = z)

#> # tidytable [10 × 3]
#>        x     y z    
#>    <int> <dbl> <chr>
#>  1    NA     1 a    
#>  2    NA     2 a    
#>  3    NA     3 a    
#>  4     4     4 a    
#>  5     5     5 a    
#>  6     6     6 a    
#>  7     7     6 a    
#>  8     8     8 a    
#>  9     9    NA b    
#> 10    10    10 b    

test_df %>%
  fill.(x, y, .by = z, .direction = "downup")

#> # tidytable [10 × 3]
#>        x     y z    
#>    <int> <dbl> <chr>
#>  1     4     1 a    
#>  2     4     2 a    
#>  3     4     3 a    
#>  4     4     4 a    
#>  5     5     5 a    
#>  6     6     6 a    
#>  7     7     6 a    
#>  8     8     8 a    
#>  9     9    10 b    
#> 10    10    10 b