Separate a character column into multiple columns

Separates a single column into multiple columns using a user supplied separator or regex.

If a separator is not supplied one will be automatically detected.

Note: Using automatic detection or regex will be slower than simple separators such as "," or ".".

separate.(.df, col, into, sep = "[^[:alnum:]]+", remove = TRUE, ...)

Arguments

.df
A data.frame or data.table

col
The column to split into multiple columns

into
New column names to split into. A character vector.

sep
Separator to split on. Can be specified or detected automatically

remove
If TRUE, remove the input column from the output data.table

...
Further argument to pass to data.table::tstrsplit

Examples

test_df <- data.table(x = c("a", "a.b", "a.b", NA))

# "sep" can be automatically detected (slower)
test_df %>%
  separate.(x, into = c("c1", "c2"))

#> # tidytable [4 × 2]
#>   c1    c2   
#>   <chr> <chr>
#> 1 a     <NA> 
#> 2 a     b    
#> 3 a     b    
#> 4 <NA>  <NA> 

# Faster if "sep" is provided
test_df %>%
  separate.(x, into = c("c1", "c2"), sep = ".")

#> # tidytable [4 × 2]
#>   c1    c2   
#>   <chr> <chr>
#> 1 a     <NA> 
#> 2 a     b    
#> 3 a     b    
#> 4 <NA>  <NA>