Select variables with a function

This selection helper selects the variables for which a function returns `TRUE`.

where(fn)

Arguments

fn
A function that returns `TRUE` or `FALSE` (technically, a _predicate_ function). Can also be a purrr-like formula.

Examples

iris %>% select.(where(is.factor))

#>        Species
#>   1:    setosa
#>   2:    setosa
#>   3:    setosa
#>   4:    setosa
#>   5:    setosa
#>  ---          
#> 146: virginica
#> 147: virginica
#> 148: virginica
#> 149: virginica
#> 150: virginica

iris %>% select.(where(is.numeric))

#>      Sepal.Length Sepal.Width Petal.Length Petal.Width
#>   1:          5.1         3.5          1.4         0.2
#>   2:          4.9         3.0          1.4         0.2
#>   3:          4.7         3.2          1.3         0.2
#>   4:          4.6         3.1          1.5         0.2
#>   5:          5.0         3.6          1.4         0.2
#>  ---                                                  
#> 146:          6.7         3.0          5.2         2.3
#> 147:          6.3         2.5          5.0         1.9
#> 148:          6.5         3.0          5.2         2.0
#> 149:          6.2         3.4          5.4         2.3
#> 150:          5.9         3.0          5.1         1.8

## The formula shorthand
# These expressions are equivalent:

iris %>% select.(where(is.numeric))

#>      Sepal.Length Sepal.Width Petal.Length Petal.Width
#>   1:          5.1         3.5          1.4         0.2
#>   2:          4.9         3.0          1.4         0.2
#>   3:          4.7         3.2          1.3         0.2
#>   4:          4.6         3.1          1.5         0.2
#>   5:          5.0         3.6          1.4         0.2
#>  ---                                                  
#> 146:          6.7         3.0          5.2         2.3
#> 147:          6.3         2.5          5.0         1.9
#> 148:          6.5         3.0          5.2         2.0
#> 149:          6.2         3.4          5.4         2.3
#> 150:          5.9         3.0          5.1         1.8

iris %>% select.(where(function(x) is.numeric(x)))

#>      Sepal.Length Sepal.Width Petal.Length Petal.Width
#>   1:          5.1         3.5          1.4         0.2
#>   2:          4.9         3.0          1.4         0.2
#>   3:          4.7         3.2          1.3         0.2
#>   4:          4.6         3.1          1.5         0.2
#>   5:          5.0         3.6          1.4         0.2
#>  ---                                                  
#> 146:          6.7         3.0          5.2         2.3
#> 147:          6.3         2.5          5.0         1.9
#> 148:          6.5         3.0          5.2         2.0
#> 149:          6.2         3.4          5.4         2.3
#> 150:          5.9         3.0          5.1         1.8

iris %>% select.(where(~ is.numeric(.x)))

#>      Sepal.Length Sepal.Width Petal.Length Petal.Width
#>   1:          5.1         3.5          1.4         0.2
#>   2:          4.9         3.0          1.4         0.2
#>   3:          4.7         3.2          1.3         0.2
#>   4:          4.6         3.1          1.5         0.2
#>   5:          5.0         3.6          1.4         0.2
#>  ---                                                  
#> 146:          6.7         3.0          5.2         2.3
#> 147:          6.3         2.5          5.0         1.9
#> 148:          6.5         3.0          5.2         2.0
#> 149:          6.2         3.4          5.4         2.3
#> 150:          5.9         3.0          5.1         1.8

# This shortahand is useful for adding logic inline.
# Here we select all numeric variables whose mean is greater than 3.5:
iris %>%
  select.(where(~ is.numeric(.x) && mean(.x) > 3.5))

#>      Sepal.Length Petal.Length
#>   1:          5.1          1.4
#>   2:          4.9          1.4
#>   3:          4.7          1.3
#>   4:          4.6          1.5
#>   5:          5.0          1.4
#>  ---                          
#> 146:          6.7          5.2
#> 147:          6.3          5.0
#> 148:          6.5          5.2
#> 149:          6.2          5.4
#> 150:          5.9          5.1