Convert character and factor columns to dummy variables

Convert character and factor columns to dummy variables

get_dummies.(
  .df,
  cols = c(where(is.character), where(is.factor)),
  prefix = TRUE,
  prefix_sep = "_",
  drop_first = FALSE,
  dummify_na = TRUE
)

Arguments

.df
A data.frame or data.table

cols
A single column or a vector of unquoted columns to dummify. Defaults to all character & factor columns using c(is.character, is.factor). tidyselect compatible.

prefix
TRUE/FALSE - If TRUE, a prefix will be added to new column names

prefix_sep
Separator for new column names

drop_first
TRUE/FALSE - If TRUE, the first dummy column will be dropped

dummify_na
TRUE/FALSE - If TRUE, NAs will also get dummy columns

Examples

test_df <- tidytable(
  col1 = c("a", "b", "c", NA),
  col2 = as.factor(c("a", "b", NA, "d")),
  var1 = rnorm(4,0,1))

# Automatically does all character/factor columns
test_df %>%
  get_dummies.()

#> # tidytable [4 × 11]
#>   col1  col2      var1 col1_a col1_b col1_c col1_NA col2_a col2_b col2_NA col2_d
#>   <chr> <fct>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#> 1 a     a     -1.40         1      0      0       0      1      0       0      0
#> 2 b     b      0.255        0      1      0       0      0      1       0      0
#> 3 c     NA    -2.44         0      0      1       0      0      0       1      0
#> 4 NA    d     -0.00557      0      0      0       1      0      0       0      1

# Can select one column
test_df %>%
  get_dummies.(col1)

#> # tidytable [4 × 7]
#>   col1  col2      var1 col1_a col1_b col1_c col1_NA
#>   <chr> <fct>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#> 1 a     a     -1.40         1      0      0       0
#> 2 b     b      0.255        0      1      0       0
#> 3 c     NA    -2.44         0      0      1       0
#> 4 NA    d     -0.00557      0      0      0       1

# Can select one or multiple columns in a vector of unquoted column names
test_df %>%
  get_dummies.(c(col1, col2))

#> # tidytable [4 × 11]
#>   col1  col2      var1 col1_a col1_b col1_c col1_NA col2_a col2_b col2_NA col2_d
#>   <chr> <fct>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#> 1 a     a     -1.40         1      0      0       0      1      0       0      0
#> 2 b     b      0.255        0      1      0       0      0      1       0      0
#> 3 c     NA    -2.44         0      0      1       0      0      0       1      0
#> 4 NA    d     -0.00557      0      0      0       1      0      0       0      1

# Can drop certain columns using
test_df %>%
  get_dummies.(c(where(is.character), -col2))

#> # tidytable [4 × 7]
#>   col1  col2      var1 col1_a col1_b col1_c col1_NA
#>   <chr> <fct>    <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#> 1 a     a     -1.40         1      0      0       0
#> 2 b     b      0.255        0      1      0       0
#> 3 c     NA    -2.44         0      0      1       0
#> 4 NA    d     -0.00557      0      0      0       1

test_df %>%
  get_dummies.(prefix_sep = ".", drop_first = TRUE)

#> # tidytable [4 × 9]
#>   col1  col2      var1 col1.b col1.c col1.NA col2.b col2.NA col2.d
#>   <chr> <fct>    <dbl>  <dbl>  <dbl>   <dbl>  <dbl>   <dbl>  <dbl>
#> 1 a     a     -1.40         0      0       0      0       0      0
#> 2 b     b      0.255        1      0       0      1       0      0
#> 3 c     NA    -2.44         0      1       0      0       1      0
#> 4 NA    d     -0.00557      0      0       1      0       0      1

test_df %>%
  get_dummies.(c(col1, col2), dummify_na = FALSE)

#> # tidytable [4 × 9]
#>   col1  col2      var1 col1_a col1_b col1_c col2_a col2_b col2_d
#>   <chr> <fct>    <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1 a     a     -1.40         1      0      0      1      0      0
#> 2 b     b      0.255        0      1      0      0      1      0
#> 3 c     NA    -2.44         0      0      1      0      0      0
#> 4 NA    d     -0.00557      0      0      0      0      0      1