Convert character and factor columns to dummy variables

Convert character and factor columns to dummy variables

Supports enhanced selection

get_dummies.(
  .data,
  cols = NULL,
  prefix = TRUE,
  prefix_sep = "_",
  drop_first = FALSE,
  dummify_na = TRUE
)

dt_get_dummies(
  .data,
  cols = NULL,
  prefix = TRUE,
  prefix_sep = "_",
  drop_first = FALSE,
  dummify_na = TRUE
)

Arguments

.data
A data.frame or data.table

cols
If NULL makes dummy variables for all character & factor columns. Or user can input a single column or a vector of unquoted columns to dummify

prefix
TRUE/FALSE - If TRUE, a prefix will be added to new column names

prefix_sep
Separator for new column names

drop_first
TRUE/FALSE - If TRUE, the first dummy column will be dropped

dummify_na
TRUE/FALSE - If TRUE, NAs will also get dummy columns

Examples

test_df <- data.table::data.table(
  col1 = c(letters[1:3], NA),
  col2 = as.factor(c(letters[3:1], NA)),
  var1 = rnorm(7,0,1))

#> Warning: Item 1 has 4 rows but longest item has 7; recycled with remainder.

#> Warning: Item 2 has 4 rows but longest item has 7; recycled with remainder.

# Automatically does all character/factor columns
test_df %>%
  get_dummies.()

#>     col1  col2         var1 col1_a col1_b col1_c col1_NA col2_c col2_b
#>    <chr> <fct>        <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>
#> 1:     a     c -1.400043517      1      0      0       0      1      0
#> 2:     b     b  0.255317055      0      1      0       0      0      1
#> 3:     c     a -2.437263611      0      0      1       0      0      0
#> 4:  <NA>  <NA> -0.005571287      0      0      0       1      0      0
#> 5:     a     c  0.621552721      1      0      0       0      1      0
#> 6:     b     b  1.148411606      0      1      0       0      0      1
#> 7:     c     a -1.821817661      0      0      1       0      0      0
#> 2 variables not shown: col2_a <dbl>, col2_NA <dbl>

# Can select one column
test_df %>%
  get_dummies.(col1)

#>     col1  col2         var1 col1_a col1_b col1_c col1_NA
#>    <chr> <fct>        <dbl>  <dbl>  <dbl>  <dbl>   <dbl>
#> 1:     a     c -1.400043517      1      0      0       0
#> 2:     b     b  0.255317055      0      1      0       0
#> 3:     c     a -2.437263611      0      0      1       0
#> 4:  <NA>  <NA> -0.005571287      0      0      0       1
#> 5:     a     c  0.621552721      1      0      0       0
#> 6:     b     b  1.148411606      0      1      0       0
#> 7:     c     a -1.821817661      0      0      1       0

# Can select one or multiple columns in a vector of unquoted column names
test_df %>%
  get_dummies.(c(col1, col2))

#>     col1  col2         var1 col1_a col1_b col1_c col1_NA col2_c col2_b
#>    <chr> <fct>        <dbl>  <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>
#> 1:     a     c -1.400043517      1      0      0       0      1      0
#> 2:     b     b  0.255317055      0      1      0       0      0      1
#> 3:     c     a -2.437263611      0      0      1       0      0      0
#> 4:  <NA>  <NA> -0.005571287      0      0      0       1      0      0
#> 5:     a     c  0.621552721      1      0      0       0      1      0
#> 6:     b     b  1.148411606      0      1      0       0      0      1
#> 7:     c     a -1.821817661      0      0      1       0      0      0
#> 2 variables not shown: col2_a <dbl>, col2_NA <dbl>

test_df %>%
  get_dummies.(prefix_sep = ".", drop_first = TRUE)

#>     col1  col2         var1 col1.b col1.c col1.NA col2.b col2.a col2.NA
#>    <chr> <fct>        <dbl>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>   <dbl>
#> 1:     a     c -1.400043517      0      0       0      0      0       0
#> 2:     b     b  0.255317055      1      0       0      1      0       0
#> 3:     c     a -2.437263611      0      1       0      0      1       0
#> 4:  <NA>  <NA> -0.005571287      0      0       1      0      0       1
#> 5:     a     c  0.621552721      0      0       0      0      0       0
#> 6:     b     b  1.148411606      1      0       0      1      0       0
#> 7:     c     a -1.821817661      0      1       0      0      1       0

test_df %>%
  get_dummies.(c(col1, col2), dummify_na = FALSE)

#>     col1  col2         var1 col1_a col1_b col1_c col2_c col2_b col2_a
#>    <chr> <fct>        <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1:     a     c -1.400043517      1      0      0      1      0      0
#> 2:     b     b  0.255317055      0      1      0      0      1      0
#> 3:     c     a -2.437263611      0      0      1      0      0      1
#> 4:  <NA>  <NA> -0.005571287      0      0      0      0      0      0
#> 5:     a     c  0.621552721      1      0      0      1      0      0
#> 6:     b     b  1.148411606      0      1      0      0      1      0
#> 7:     c     a -1.821817661      0      0      1      0      0      1