Summarize multiple columns

Summarize multiple columns simultaneously

summarize_across.(
  .df,
  .cols = everything(),
  .fns = NULL,
  ...,
  .by = NULL,
  .names = NULL
)

summarise_across.(
  .df,
  .cols = everything(),
  .fns = NULL,
  ...,
  .by = NULL,
  .names = NULL
)

Arguments

.df

A data.frame or data.table

.cols

vector c() of unquoted column names. tidyselect compatible.

.fns

Functions to pass. Can pass a list of functions.

...

Other arguments for the passed function

.by

Columns to group by

.names

A glue specification that helps with renaming output columns. {.col} stands for the selected column, and {.fn} stands for the name of the function being applied. The default (NULL) is equivalent to "{.col}" for a single function case and "{.col}_{.fn}" when a list is used for .fns.

Examples

test_df <- data.table(
  a = 1:3,
  b = 4:6,
  z = c("a", "a", "b")
)

# Pass a single function
test_df %>%
  summarize_across.(c(a, b), mean, na.rm = TRUE)
#> Warning: `summarize_across.()` was deprecated in tidytable 0.6.4.
#> Please use `summarize.(across.())
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#> # A tidytable: 1 × 2
#>       a     b
#>   <dbl> <dbl>
#> 1     2     5

# Single function using purrr style interface
test_df %>%
  summarize_across.(c(a, b), ~ mean(.x, na.rm = TRUE))
#> # A tidytable: 1 × 2
#>       a     b
#>   <dbl> <dbl>
#> 1     2     5

# Passing a list of functions (with .by)
test_df %>%
  summarize_across.(c(a, b), list(mean, max), .by = z)
#> # A tidytable: 2 × 5
#>   z       a_1   a_2   b_1   b_2
#>   <chr> <dbl> <int> <dbl> <int>
#> 1 a       1.5     2   4.5     5
#> 2 b       3       3   6       6

# Passing a named list of functions (with .by)
test_df %>%
  summarize_across.(c(a, b),
                    list(avg = mean,
                         max = ~ max(.x)),
                    .by = z)
#> # A tidytable: 2 × 5
#>   z     a_avg a_max b_avg b_max
#>   <chr> <dbl> <int> <dbl> <int>
#> 1 a       1.5     2   4.5     5
#> 2 b       3       3   6       6

# Use the `.names` argument for more naming control
test_df %>%
  summarize_across.(c(a, b),
                    list(avg = mean,
                         max = ~ max(.x)),
                    .by = z,
                    .names = "{.col}_test_{.fn}")
#> # A tidytable: 2 × 5
#>   z     a_test_avg a_test_max b_test_avg b_test_max
#>   <chr>      <dbl>      <int>      <dbl>      <int>
#> 1 a            1.5          2        4.5          5
#> 2 b            3            3        6            6