Unnest a nested data.table

Unnest a nested data.table.

unnest.(.df, ..., .drop = TRUE, names_sep = NULL, names_repair = "unique")

Arguments

.df
A nested data.table

...
Columns to unnest. If empty, unnests all list columns. tidyselect compatible.

.drop
Should list columns that were not unnested be dropped

names_sep
If NULL, the default, the inner column names will become the new outer column names.

If a string, the name of the outer column will be appended to the beginning of the inner column names, with names_sep used as a separator.

names_repair
Treatment of duplicate names. See ?vctrs::vec_as_names for options/details.

Examples

nested_df <-
  data.table(
    a = 1:10,
    b = 11:20,
    c = c(rep("a", 6), rep("b", 4)),
    d = c(rep("a", 4), rep("b", 6))
  ) %>%
  nest_by.(c, d) %>%
  mutate.(pulled_vec = map.(data, ~ pull.(.x, a)))

nested_df %>%
  unnest.(data)

#> # A tidytable: 10 × 4
#>    c     d         a     b
#>    <chr> <chr> <int> <int>
#>  1 a     a         1    11
#>  2 a     a         2    12
#>  3 a     a         3    13
#>  4 a     a         4    14
#>  5 a     b         5    15
#>  6 a     b         6    16
#>  7 b     b         7    17
#>  8 b     b         8    18
#>  9 b     b         9    19
#> 10 b     b        10    20

nested_df %>%
  unnest.(data, names_sep = "_")

#> # A tidytable: 10 × 4
#>    c     d     data_a data_b
#>    <chr> <chr>  <int>  <int>
#>  1 a     a          1     11
#>  2 a     a          2     12
#>  3 a     a          3     13
#>  4 a     a          4     14
#>  5 a     b          5     15
#>  6 a     b          6     16
#>  7 b     b          7     17
#>  8 b     b          8     18
#>  9 b     b          9     19
#> 10 b     b         10     20

nested_df %>%
  unnest.(data, pulled_vec)

#> # A tidytable: 10 × 5
#>    c     d         a     b pulled_vec
#>    <chr> <chr> <int> <int>      <int>
#>  1 a     a         1    11          1
#>  2 a     a         2    12          2
#>  3 a     a         3    13          3
#>  4 a     a         4    14          4
#>  5 a     b         5    15          5
#>  6 a     b         6    16          6
#>  7 b     b         7    17          7
#>  8 b     b         8    18          8
#>  9 b     b         9    19          9
#> 10 b     b        10    20         10